Change quantity to float

+++ src/main/java/fr/jahland/baby/model/DrugInfo.java
@@ -5,5 +5,5 @@ package fr.jahland.baby.model;
  */
 public class DrugInfo {
     public Masterdata drugType;
-    public Integer quantity;
+    public Float quantity;
 }
